jre pipeline upgrade

+++ src_re/main.py
@@ -114,17 +114,20 @@ if __name__ == "__main__":
     parser.add_argument('--cache_dir', type=str, default="/blue/woodard/share/Relation-Extraction/LLM_for_RE/cache", help="LLM cache directory")
 
     parser.add_argument("--config_file", type=str, default=None,
-                        help="path to config file", required=True)
+                        help="path to config file", required=False)
     parser.add_argument('--redo', type=bool, default=False)
     args = parser.parse_args()
 
     if args.config_file:
+        config_file = args.config_file
         with open(args.config_file, 'r') as f:
             args.__dict__ = json.load(f)
+            setattr(args, 'config_file', config_file)
 
     try:
         main(args)
-        os.remove(args.config_file)
+        if args.config_file:
+            os.remove(args.config_file)
     except Exception as e:
         print(f'[Error] {e}')
         print(traceback.format_exc())
@@ -133,6 +136,5 @@ if __name__ == "__main__":
         os.makedirs(redo_bin, exist_ok=True)
         with open(f'{redo_bin}/exp-{args.demo}_{args.prompt}_{args.k}.json', 'w') as f:
             json.dump(args.__dict__, f)
-
 
     print('\tDone.')
